🩹 update tests

+++ memcache/memcache.go
@@ -71,9 +71,10 @@ func (s *Storage) Get(key string) ([]byte, error) {
 }
 
 // Set key with value
+// Set key with value
 func (s *Storage) Set(key string, val []byte, exp time.Duration) error {
 	// Ain't Nobody Got Time For That
-	if len(val) <= 0 {
+	if len(key) <= 0 || len(val) <= 0 {
 		return nil
 	}
 	item := s.acquireItem()
@@ -90,6 +91,10 @@ func (s *Storage) Set(key string, val []byte, exp time.Duration) error {
 
 // Delete key by key
 func (s *Storage) Delete(key string) error {
+	// Ain't Nobody Got Time For That
+	if len(key) <= 0 {
+		return nil
+	}
 	return s.db.Delete(key)
 }
